Добавил представление для Node Доработка нод

+++ BoardGame/Model/Map/Node.cs
@@ -11,7 +11,7 @@ namespace BoardGame.Model.Map
     {
         public CoordinatePoint Coordinate { get; set; }
         public List<Node> Neighbors { get; set; }
-        public List<int> Side { get; set; }
+        public List<double> Side { get; set; }
         public DifficultyAreaType DifficultyAreaType { get; set; }
     }
 }
